+++ WLUWEngine/WLUWEngine/src/Header.h
@@ -2,4 +2,4 @@
 
 #include <cmath>
 
-const double epsilon = pow(1.0, -5.0);
+const float epsilon = pow(1.0, -5.0);
